app/requirements.txt
Django==4.2
djangorestframework==3.14.0
drf-spectacular==0.26.1
psycopg2-binary==2.9.1
PyJWT==2.6.0
django-cors-headers==3.14.0
isort==5.11.3
flake8==6.1.0
python-decouple==3.7
drf-yasg
django-cors-headers==3.14.0
reportlab